payment/requirements.txt
Flask==2.3.1
gunicorn==20.1.0
redis==4.5.4
python-dotenv==1.0.0
requests
google-cloud-spanner==3.35.0
kafka-python
sqlalchemy
sqlalchemy-spanner